circuit: merge email verifier tests to one file
