--- packages/circuits/tests/no-body-hash.test.ts
@@ -1,54 +1,0 @@
-import { DKIMVerificationResult } from "@zk-email/helpers/src/dkim";
-import { generateCircuitInputs } from "@zk-email/helpers/src/input-helpers";
-
-const { verifyDKIMSignature } = require("@zk-email/helpers/src/dkim");
-const fs = require("fs");
-const path = require("path");
-const wasm_tester = require("circom_tester").wasm;
-const F1Field = require("ffjavascript").F1Field;
-const Scalar = require("ffjavascript").Scalar;
-
-exports.p = Scalar.fromString(
-  "21888242871839275222246405745257275088548364400416034343698204186575808495617"
-);
-
-describe("EmailVerifier : Without body check", () => {
-  jest.setTimeout(10 * 60 * 1000); // 10 minutes
-
-  let dkimResult: DKIMVerificationResult;
-  let circuit: any;
-
-  beforeAll(async () => {
-    const rawEmail = fs.readFileSync(
-      path.join(__dirname, "./test.eml"),
-      "utf8"
-    );
-    dkimResult = await verifyDKIMSignature(rawEmail);
-
-    circuit = await wasm_tester(
-      path.join(__dirname, "./test-circuits/no-body-hash.test.circom"),
-      {
-        recompile: true,
-        include: path.join(__dirname, "../../../node_modules"),
-        output: path.join(__dirname, "./compiled-test-circuits"),
-      }
-    );
-  });
-
-  it("should verify email when ignore_body_hash_check is true", async function () {
-    // The result wont have shaPrecomputeSelector, maxMessageLength, maxBodyLength, ignoreBodyHashCheck
-    const emailVerifierInputs = generateCircuitInputs({
-      rsaSignature: dkimResult.signature,
-      rsaPublicKey: dkimResult.publicKey,
-      body: dkimResult.body,
-      bodyHash: dkimResult.bodyHash,
-      message: dkimResult.message,
-      maxMessageLength: 640,
-      maxBodyLength: 768,
-      ignoreBodyHashCheck: true,
-    });
-
-    const witness = await circuit.calculateWitness(emailVerifierInputs);
-    await circuit.checkConstraints(witness);
-  });
-});
